Allow compile-time defines to overrule ffconf.h
All settings in ffconf.h can be overwritten by setting a compile-time
define in the top level CMakeLists.txt.
For example, a project might add the following to its CMakeLists.txt:

add_definitions(-D_FS_READONLY=1)

To overrule the ffconf.h settings and build a readonly version of the
library.